misys_bridge_server/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pymssql>=2.2.0
python-dotenv>=1.0.0
